# docs/open-api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "DiffScope Synthesis Platform API",
    "version": "0.0.0",
    "description": "DSSP HTTP API surface represented by this service. A server implementation can support a subset of the broader DSSP API."
  },
  "servers": [
    {
      "url": "http://localhost:13711",
      "description": "Default local server URL. The actual host and port are configured by the service."
    }
  ],
...
  "paths": {
...
    "/api/arch/{arch_id}": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchitecture",
        "summary": "Get architecture metadata",
        "description": "Returns metadata for one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Architecture"
          },
          "404": {
            "$ref": "#/components/responses/ArchitectureNotFound"
          },
          "422": {
            "$ref": "#/components/responses/ArchitectureOperationError"
          }
        }
      }
    },
    "/api/singer": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getSingerList",
        "summary": "List all singers",
        "description": "Returns singers from every registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SingerList"
          },
          "404": {
            "$ref": "#/components/responses/SingerNotFound"
          },
          "422": {
            "$ref": "#/components/responses/SingerOperationError"
          }
        }
      }
    },
    "/api/arch/{arch_id}/singer": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSingerList",
        "summary": "List singers for an architecture",
        "description": "Returns singers from one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ArchSingerList"
          },
          "404": {
            "$ref": "#/components/responses/SingerNotFound"
          },
          "422": {
            "$ref": "#/components/responses/SingerOperationError"
          }
        }
      }
    },
    "/api/arch/{arch_id}/singer/{singer_id}": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSinger",
        "summary": "Get singer metadata",
        "description": "Returns one singer from one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/SingerID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Singer"
          },
          "404": {
            "$ref": "#/components/responses/SingerNotFound"
          },
          "422": {
            "$ref": "#/components/responses/SingerOperationError"
          }
        }
      }
    },
    "/api/arch/{arch_id}/singer/{singer_id}/avatar": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSingerAvatar",
        "summary": "Get singer avatar URL",
        "description": "Returns the avatar URL for one singer.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/SingerID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SingerAvatar"
          },
          "404": {
            "$ref": "#/components/responses/SingerNotFound"
          },
          "422": {
            "$ref": "#/components/responses/SingerOperationError"
          }
        }
      }
    },
    "/api/arch/{arch_id}/singer/{singer_id}/background": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSingerBackground",
        "summary": "Get singer background URL",
        "description": "Returns the background image URL for one singer.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/SingerID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SingerBackground"
          },
          "404": {
            "$ref": "#/components/responses/SingerNotFound"
          },
          "422": {
            "$ref": "#/components/responses/SingerOperationError"
          }
        }
      }
    },
    "/api/arch/{arch_id}/singer/{singer_id}/demo_audio": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSingerDemoAudioList",
        "summary": "List singer demo audio",
        "description": "Returns demo audio entries for one singer.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/SingerID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SingerDemoAudioList"
          },
          "404": {
            "$ref": "#/components/responses/SingerNotFound"
          },
          "422": {
            "$ref": "#/components/responses/SingerOperationError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ArchID": {
        "name": "arch_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Architecture identifier."
      },
      "SingerID": {
        "name": "singer_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Singer identifier understood by the selected architecture."
      },
      "DisplayLanguage": {
        "name": "display_language",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Preferred language for display metadata."
      }
    },
...
    "responses": {
...
      "Architecture": {
        "description": "Architecture metadata.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureMetadata"
            }
          }
        }
      },
      "ArchitectureNotFound": {
        "description": "The requested architecture does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureNotFoundAPIError"
            }
          }
        }
      },
      "ArchitectureOperationError": {
        "description": "The architecture metadata request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureOperationAPIError"
            }
          }
        }
      },
      "SingerList": {
        "description": "Singer metadata from every registered architecture.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerInfoWithArchList"
            }
          }
        }
      },
      "ArchSingerList": {
        "description": "Singer metadata from one registered architecture.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerInfoList"
            }
          }
        }
      },
      "Singer": {
        "description": "Singer metadata.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerInfo"
            }
          }
        }
      },
      "SingerAvatar": {
        "description": "Singer avatar URL.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerAvatarResponse"
            }
          }
        }
      },
      "SingerBackground": {
        "description": "Singer background URL.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerBackgroundResponse"
            }
          }
        }
      },
      "SingerDemoAudioList": {
        "description": "Singer demo audio entries.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerDemoAudioList"
            }
          }
        }
      },
      "SingerNotFound": {
        "description": "The requested architecture or singer does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerNotFoundAPIError"
            }
          }
        }
      },
      "SingerOperationError": {
        "description": "The singer metadata request was syntactically valid but could not be completed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerOperationAPIError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ArchitectureMetadata": {
        "type": "object",
        "required": [
          "id",
          "name",
          "pronunciation_mode",
          "phoneme_mode",
          "parameters",
          "audio_dependencies"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "name": {
            "type": "string"
          },
          "pronunciation_mode": {
            "$ref": "#/components/schemas/PronunciationMode"
          },
          "phoneme_mode": {
            "$ref": "#/components/schemas/PhonemeMode"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ArchitectureParameterMetadata"
            }
          },
          "audio_dependencies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "PronunciationMode": {
        "type": "string",
        "enum": [
          "FULL",
          "SKIP"
        ]
      },
      "PhonemeMode": {
        "type": "string",
        "enum": [
          "FULL",
          "TOKEN_ONLY",
          "SKIP"
        ]
      },
      "ArchitectureParameterMetadata": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/DirectArchitectureParameterMetadata"
          },
          {
            "$ref": "#/components/schemas/IndirectArchitectureParameterMetadata"
          }
        ]
      },
      "DirectArchitectureParameterMetadata": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "const": "DIRECT"
          }
        },
        "additionalProperties": false
      },
      "IndirectArchitectureParameterMetadata": {
        "type": "object",
        "required": [
          "type",
          "depends_on"
        ],
        "properties": {
          "type": {
            "const": "INDIRECT"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "SingerInfoWithArchList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SingerInfoWithArch"
        }
      },
      "SingerInfoList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SingerInfo"
        }
      },
      "SingerInfoWithArch": {
        "type": "object",
        "required": [
          "id",
          "name",
          "arch",
          "languages",
          "default_language",
          "extra",
          "default_extra"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Singer identifier understood by the selected architecture."
          },
          "name": {
            "type": "string"
          },
          "arch": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "default_language": {
            "type": "string"
          },
          "extra": {
            "description": "Architecture-specific singer JSON payload."
          },
          "default_extra": {
            "description": "Default architecture-specific singer JSON payload."
          }
        },
        "additionalProperties": false
      },
      "SingerInfo": {
        "type": "object",
        "required": [
          "id",
          "name",
          "languages",
          "default_language",
          "extra",
          "default_extra"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Singer identifier understood by the selected architecture."
          },
          "name": {
            "type": "string"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "default_language": {
            "type": "string"
          },
          "extra": {
            "description": "Architecture-specific singer JSON payload."
          },
          "default_extra": {
            "description": "Default architecture-specific singer JSON payload."
          }
        },
        "additionalProperties": false
      },
      "SingerAvatarResponse": {
        "type": "object",
        "required": [
          "avatar_url"
        ],
        "properties": {
          "avatar_url": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SingerBackgroundResponse": {
        "type": "object",
        "required": [
          "background_url"
        ],
        "properties": {
          "background_url": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SingerDemoAudioList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SingerDemoAudio"
        }
      },
      "SingerDemoAudio": {
        "type": "object",
        "required": [
          "name",
          "audio_url"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "audio_url": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ArchitectureNotFoundAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/ArchitectureNotFoundErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ArchitectureOperationAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/ArchitectureOperationErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SingerNotFoundAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/SingerNotFoundErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SingerOperationAPIError": {
        "type": "object",
        "required": [
          "state",
          "code",
          "message"
        ],
        "properties": {
          "state": {
            "const": "ERROR"
          },
          "code": {
            "$ref": "#/components/schemas/SingerOperationErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ArchitectureNotFoundErrorCode": {
        "type": "string",
        "enum": [
          "UNKNOWN_ARCH"
        ]
      },
      "ArchitectureOperationErrorCode": {
        "type": "string",
        "enum": [
          "INTERNAL_ERROR"
        ]
      },
      "SingerNotFoundErrorCode": {
        "type": "string",
        "enum": [
          "UNKNOWN_ARCH",
          "SINGER_NOT_EXIST"
        ]
      },
      "SingerOperationErrorCode": {
        "type": "string",
        "enum": [
          "INTERNAL_ERROR"
        ]
      },
...
    }
  }
}